S3 functionality unit tests. Reorg of all other unit tests for sanity purposes.
