+++ .github/workflows/python-coverage.yaml
@@ -11,7 +11,6 @@ on:
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     steps:
@@ -26,6 +25,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r DEV_REQUIREMENTS.txt
+        pip install -r S3_REQUIREMENTS.txt
 
     - name: Test with pytest
       run: |
