+++ .github/workflows/python-pytest-s3.yaml
@@ -1,7 +1,7 @@
-# This workflow will install Python dependencies, run general tests with a variety of Python versions
+# This workflow will install Python dependencies, run S3 tests with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: pytest
+name: pytest-s3
 
 on:
   pull_request:
@@ -28,10 +28,11 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e .
         pip install -r DEV_REQUIREMENTS.txt
+        pip install -r S3_REQUIREMENTS.txt
 
     - name: Test with pytest
       run: |
-        pytest tests/base --cov=spock --cov-config=.coveragerc --junitxml=junit/test-results-${{ matrix.python-version }}.xml --cov-report=xml --cov-report=html
+        pytest tests/s3 --cov=spock --cov-config=.coveragerc --junitxml=junit/test-results-${{ matrix.python-version }}.xml --cov-report=xml --cov-report=html
 
     - name: Upload pytest test results
       uses: actions/upload-artifact@v2
