+++ S3_REQUIREMENTS.txt
@@ -1,0 +1,4 @@
+boto3
+botocore
+hurry.filesize
+s3transfer
