+++ spock/addons/s3/configs.py
@@ -49,16 +49,20 @@ class S3Config:
     *Attributes*:
 
         session: instantiated boto3 session object
-        s3_session: automatically generated s3 client from the boto3 session
+        s3_session: automatically generated s3 client from the boto3 session if not provided
         kms_arn: AWS KMS key ARN (optional)
         temp_folder: temporary working folder to write/read spock configuration(s) (optional: defaults to /tmp)
+        download_config: S3DownloadConfig for extra download configs (optional)
+        upload_config: S3UploadConfig for extra upload configs (optional)
 
     """
     session: boto3.Session
-    s3_session: BaseClient = attr.ib(init=False)
+    # s3_session: BaseClient = attr.ib(init=False)
+    s3_session: typing.Optional[BaseClient] = None
     temp_folder: typing.Optional[str] = '/tmp/'
     download_config: S3DownloadConfig = S3DownloadConfig()
     upload_config: S3UploadConfig = S3UploadConfig()
 
     def __attrs_post_init__(self):
-        self.s3_session = self.session.client('s3')
+        if self.s3_session is None:
+            self.s3_session = self.session.client('s3')
